♻️ refactor: extract demo chat-stream to sibling file
Move chatStream / promoCardInset / agentPosition out of
ModelDownloadHostView.swift (396 → 315 lines) into a new
ModelDownloadHostView+ChatStream.swift sibling, matching the existing
+GameHeader / +StateFallbacks / +ControlBar / +Routing split. Creates
headroom under swiftlint's 400-line file_length cap for the upcoming
scroll-follow wiring (#785).

Pure move, no behavior change. Widens reduceMotion / currentState /
replayHadStarted / sources from private to internal so the sibling
extension can read them (this file's existing "drop private so the
sibling can read it" convention).

See #785
