+++ Pastura/Pastura/Views/ModelDownload/ModelDownloadHostView.swift
@@ -78,11 +78,18 @@ struct ModelDownloadHostView: View {
   static let minPlayableDemoCount = 2
 
   @Environment(\.scenePhase) private var scenePhase
-  @Environment(\.accessibilityReduceMotion) private var reduceMotion
+  // Module-internal (not `private`) so the sibling `+ChatStream.swift`
+  // extension can read it for the row `.transition` + scroll-follow
+  // animation gating. `private` only reaches same-file extensions.
+  @Environment(\.accessibilityReduceMotion) var reduceMotion
 
   @State private var replayVM: ReplayViewModel?
-  @State private var replayHadStarted: Bool = false
-  @State private var sources: [any ReplaySource] = []
+  // Internal (not `private`) so `+ChatStream.swift`'s `promoCardInset`
+  // can read it. Same rationale as `reduceMotion` above.
+  @State var replayHadStarted: Bool = false
+  // Internal (not `private`) so `+ChatStream.swift`'s `agentPosition`
+  // / `promoCardInset` and `currentPresetName` can read it.
+  @State var sources: [any ReplaySource] = []
   @State private var isShowingCancelConfirmation: Bool = false
   // Thought-visibility (`▸ THINKING` expanded) now lives on `ReplayViewModel`
   // (`showAllThoughts`) so the pacing floor can read it; the control bar binds
@@ -101,7 +108,10 @@ struct ModelDownloadHostView: View {
 
   /// Per-descriptor download state. Defaults to `.checking` if the entry is
   /// missing from the state dict (only expected pre-`checkModelStatus`).
-  private var currentState: ModelState {
+  ///
+  /// Module-internal (not `private`) so `+ChatStream.swift`'s
+  /// `promoCardInset` can read it.
+  var currentState: ModelState {
     modelManager.state[descriptor.id] ?? .checking
   }
 
@@ -206,84 +216,6 @@ struct ModelDownloadHostView: View {
     onReady?(modelPath)
   }
 
-  private func chatStream(viewModel: ReplayViewModel) -> some View {
-    VStack(spacing: 0) {
-      gameHeader(viewModel: viewModel)
-
-      ScrollViewReader { proxy in
-        ScrollView {
-          // `spacing` uses the ChatBubbleLayout.bubbleSpacing token so a
-          // future design-system tweak flows through Demo / Sim / Results
-          // in one place. Production value is 8pt project-wide (#273 PR 2);
-          // see the token's docstring for the historical 14pt prototype
-          // reference and the divergence rationale.
-          LazyVStack(alignment: .leading, spacing: ChatBubbleLayout.bubbleSpacing) {
-            // `isLatest` keys off the last agent output, not the last
-            // chatItem (#208), so a `.demoBoundary` does not retrigger
-            // the typing animation on the bubble that precedes it.
-            let lastAgentId = viewModel.agentOutputs.last?.id
-            ForEach(viewModel.chatItems) { item in
-              switch item {
-              case .agentOutput(let entry):
-                AgentOutputRow(
-                  agent: entry.agent,
-                  output: entry.output,
-                  phaseType: entry.phaseType,
-                  showAllThoughts: viewModel.showAllThoughts,
-                  isLatest: entry.id == lastAgentId,
-                  charsPerSecond: viewModel.typingCharsPerSecond,
-                  agentPosition: agentPosition(for: entry.agent, viewModel: viewModel)
-                )
-                .id(entry.id)
-                .transition(reduceMotion ? .identity : .opacity)
-              case .demoBoundary(let id, let scenarioName):
-                DemoBoundaryRow(scenarioName: scenarioName)
-                  .id(id)
-                  .transition(reduceMotion ? .identity : .opacity)
-              }
-            }
-          }
-          // Screen-level gutters (20pt horizontal / 8pt top) match the
-          // reference HTML `.stream { padding: 8px 20px 16px }`. Intentional
-          // literals — these are container-level, not per-bubble.
-          .padding(.horizontal, 20)
-          .padding(.top, 8)
-          .animation(
-            reduceMotion ? nil : .easeOut(duration: 0.7),
-            value: viewModel.chatItems.count)
-        }
-        .onChange(of: viewModel.chatItems.count) { _, _ in
-          guard let lastId = viewModel.chatItems.last?.id else { return }
-          withAnimation(reduceMotion ? nil : .easeOut(duration: 0.3)) {
-            proxy.scrollTo(lastId, anchor: .bottom)
-          }
-        }
-      }
-
-      // Sim-style frosted controlBar (#273): mirrors `SimulationView.controlBar`
-      // shape so users learn the layout before the live simulation. Pause /
-      // Speed / Toggle interactive — Pause → `viewModel.userPause()`/
-      // `userResume()`, Speed → `viewModel.playbackSpeed` (#290).
-      controlBar(viewModel: viewModel)
-    }
-    .background(Color.screenBackground.ignoresSafeArea())
-    .safeAreaInset(edge: .bottom, spacing: Spacing.l) { promoCardInset }
-  }
-
-  /// PromoCard lives in the bottom safe area (not a ZStack overlay) so the
-  /// ScrollView viewport shrinks to exclude the card's footprint;
-  /// `scrollTo(lastId, anchor: .bottom)` then lands the newest message
-  /// above the card. The earlier `.padding(.bottom, 160)` reserved space
-  /// but didn't shrink the viewport — the anchor slid under the overlay.
-  private var promoCardInset: some View {
-    PromoCard(
-      modelState: currentState,
-      replayHadStarted: replayHadStarted,
-      totalBytes: descriptor.fileSize,
-      onRetry: { modelManager.startDownload(descriptor: descriptor) },
-      onCancel: triggerCancelConfirmation)
-  }
-
   // Module-internal so the sibling `+GameHeader.swift` extension can
   // call this helper. `private` only reaches same-file extensions.
   func currentPresetName(viewModel: ReplayViewModel) -> String {
@@ -291,20 +223,6 @@ struct ModelDownloadHostView: View {
       sourceIndex < sources.count
     else { return "" }
     return sources[sourceIndex].scenario.name
-  }
-
-  /// Agent's zero-based index in the current replay's agent list, used
-  /// by ``AvatarSlot`` for position-priority avatar color assignment.
-  /// Returns `nil` when no replay is active or the agent isn't in the
-  /// current source's `agents` list; the row then falls back to the
-  /// name-based avatar resolution.
-  private func agentPosition(
-    for agentName: String, viewModel: ReplayViewModel
-  ) -> Int? {
-    guard let sourceIndex = viewModel.currentSourceIndex,
-      sourceIndex < sources.count
-    else { return nil }
-    return sources[sourceIndex].scenario.personas.firstIndex(where: { $0.name == agentName })
   }
 
   // MARK: - Lifecycle
@@ -377,9 +295,10 @@ struct ModelDownloadHostView: View {
 
 }
 
-// `DLCompleteOverlay`, the per-state UI helpers, and the pure routing
-// functions live in their own files so this one stays under swiftlint's
-// 400-line cap. See `DLCompleteOverlay.swift`,
+// `DLCompleteOverlay`, the chat-stream rendering, the per-state UI
+// helpers, and the pure routing functions live in their own files so this
+// one stays under swiftlint's 400-line cap. See `DLCompleteOverlay.swift`,
+// `ModelDownloadHostView+ChatStream.swift`,
 // `ModelDownloadHostView+StateFallbacks.swift`, and
 // `ModelDownloadHostView+Routing.swift`.
